terraform: gsi 제거 feat: wip: 서버리스 배포 준비 refactor: 파일 분리
