+++ packages/ptt-datalake/handler/petitions.ts
@@ -1,0 +1,9 @@
+import {petitions} from '../src/petitions'
+
+export const handler = () => {
+  return petitions()
+}
+
+if (process.env.RUN) {
+  handler()
+}
