+++ packages/ptt-datalake/src/constant.ts
@@ -1,0 +1,4 @@
+export const const_min_5 = 60 * 5
+export const const_url = 'https://www1.president.go.kr'
+export const const_best_url = `${const_url}/petitions/best`
+export const const_yyyyMMDdTHHMmxx = 'yyyy-MM-dd\'T\'HH:mmxx'
