+++ packages/ptt-datalake/src/entity/createAtDocument.ts
@@ -1,0 +1,11 @@
+import {AccDocument} from '../type'
+
+export const createAtDocument = (dateX: string) =>
+  (a: AccDocument) => {
+    return {
+      hk    : a.hk,
+      rk    : `at#${dateX}`,
+      people: a.people,
+      ttl   : a.ttl,
+    }
+  }
